Seperate quotes and commands into individual files
