+++ quotes.py
@@ -1,0 +1,27 @@
+copypasta1 = """*jaw drops to floor, eyes pop out of sockets accompanied by trumpets, heart beats out of chest, awooga awooga sound effect, pulls chain on train whistle that has appeared next to head as steam blows out, slams fists on table, rattling any plates, bowls or silverware, whistles loudly, fireworks shoot from top of head, pants loudly as tongue hangs out of mouth, wipes comically large bead of sweat from forehead, clears throat, straightens tie, combs hair* Ahem, you look very lovely."""
+vgmgrules = """**-Rules-**  
+
+1. Each song has 2 points attached to it. Guessing the song title gives you 1 point, and guessing the game title gives you 1 point.
+
+2. 0.5 points will be given for partial guesses. 
+
+3. Search engines are not allowed. Hints will be given half way through the song.
+
+4. Only Western localised titles accepted. No unofficial translations. 
+
+5. Punctuation such as full stops, colons, etc... in titles do not matter.
+
+6. You can use abbreviations or shortenings for game titles as long as they are recognizable.
+
+7. Please try to guess the entire title, including its number in the series if it is part of one.
+
+8. All decisions are subject to committee discretion.
+
+9. Submit your final answers in a text document to a committee member or helper when the game is over and it will be marked!!"""
+
+lemonade ="""
+```All right, I've been thinking. When life gives you lemons? Don't make lemonade. 
+Make life take the lemons back! Get mad! 'I don't want your damn lemons! What am I supposed to do with these?`
+Demand to see life's manager! Make life rue the day it thought it could give Gurg lemons! Do you know who I am? 
+I'm the man who's going to burn your house down! With the lemons! 
+I'm going to get my engineers to invent a combustible lemon that burns your house down!```"""
